#8 [chore] modee에서 moddy로 패키지 명 및 파일 명 수정
