+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'modee-server'
+rootProject.name = 'moddy-server'
