+++ src/main/java/com/moddy/server/ServerApplication.java
@@ -1,4 +1,4 @@
-package com.modee.server;
+package com.moddy.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
