+++ src/main/java/com/moddy/server/common/dto/ErrorResponse.java
@@ -1,6 +1,6 @@
-package com.modee.server.common.dto;
+package com.moddy.server.common.dto;
 
-import com.modee.server.common.exception.enums.ErrorCode;
+import com.moddy.server.common.exception.enums.ErrorCode;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
