+++ src/main/java/com/moddy/server/common/exception/enums/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.modee.server.common.exception.enums;
+package com.moddy.server.common.exception.enums;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,8 +6,8 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor
-public enum SuccessCode {
-    ;
+public enum ErrorCode {
+    INTERNAL_SERVER_EXCEPTION(HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 오류입니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
